Implement the latest specification
